// src/api/model/insurancePolicyProductGet.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InsurancePolicyProductGet { 
    /**
     * Id of the created resource
     */
    readonly id: string;
    name: string;
    annualPremium: number;
    coverageCap: number;
    /**
     * Type of the insurance policy
     */
    type: InsurancePolicyProductGet.TypeEnum;
}
export namespace InsurancePolicyProductGet {
    export type TypeEnum = 'Car' | 'Life' | 'Home';
    export const TypeEnum = {
        Car: 'Car' as TypeEnum,
        Life: 'Life' as TypeEnum,
        Home: 'Home' as TypeEnum
    };
}